Separate inspect from other utils
