+++ src/components/form-file/form-file.js
@@ -2,12 +2,12 @@ import Vue from '../../utils/vue'
 import cloneDeep from '../../utils/clone-deep'
 import identity from '../../utils/identity'
 import looseEqual from '../../utils/loose-equal'
-import { from as arrayFrom, flatten, flattenDeep, isArray } from '../../utils/array'
+import { from as arrayFrom, flatten, flattenDeep } from '../../utils/array'
 import { getComponentConfig } from '../../utils/config'
 import { closest } from '../../utils/dom'
 import { hasPromiseSupport } from '../../utils/env'
 import { EVENT_OPTIONS_PASSIVE, eventOn, eventOff, stopEvent } from '../../utils/events'
-import { isFile, isFunction, isNull, isUndefinedOrNull } from '../../utils/inspect'
+import { isArray, isFile, isFunction, isNull, isUndefinedOrNull } from '../../utils/inspect'
 import { File } from '../../utils/safe-types'
 import { escapeRegExp } from '../../utils/string'
 import { warn } from '../../utils/warn'
