+++ src/components/form-select/form-select.js
@@ -1,7 +1,8 @@
 import Vue from '../../utils/vue'
-import { from as arrayFrom, isArray } from '../../utils/array'
+import { from as arrayFrom } from '../../utils/array'
 import { attemptBlur, attemptFocus } from '../../utils/dom'
 import { htmlOrText } from '../../utils/html'
+import { isArray } from '../../utils/inspect'
 import formCustomMixin from '../../mixins/form-custom'
 import formMixin from '../../mixins/form'
 import formSizeMixin from '../../mixins/form-size'
