+++ src/components/tooltip/tooltip.js
@@ -1,8 +1,8 @@
 import Vue from '../../utils/vue'
 import getScopId from '../../utils/get-scope-id'
-import { isArray, arrayIncludes } from '../../utils/array'
+import { arrayIncludes } from '../../utils/array'
 import { getComponentConfig } from '../../utils/config'
-import { isString, isUndefinedOrNull } from '../../utils/inspect'
+import { isArray, isString, isUndefinedOrNull } from '../../utils/inspect'
 import { HTMLElement, SVGElement } from '../../utils/safe-types'
 import { BVTooltip } from './helpers/bv-tooltip'
