+++ src/utils/inspect.js
@@ -1,5 +1,3 @@
-import { isArray } from './array'
-import { isObject, isPlainObject } from './object'
 import { File } from './safe-types'
 
 // --- Convenience inspection utilities ---
@@ -19,6 +17,17 @@ export const isEmptyString = val => val === ''
 export const isUndefinedOrNull = val => isUndefined(val) || isNull(val)
 
 export const isUndefinedOrNullOrEmpty = val => isUndefinedOrNull(val) || isEmptyString(val)
+
+export const isArray = val => Array.isArray(val)
+
+// Quick object check
+// This is primarily used to tell objects from primitive values
+// Note object could be a complex type like array, Date, etc.
+export const isObject = obj => obj !== null && typeof obj === 'object'
+
+// Strict object type check
+// Only returns `true` for plain JavaScript objects
+export const isPlainObject = obj => Object.prototype.toString.call(obj) === '[object Object]'
 
 export const isFunction = val => toType(val) === 'function'
 
@@ -40,6 +49,3 @@ export const isRegExp = val => toRawType(val) === 'RegExp'
 
 export const isPromise = val =>
   !isUndefinedOrNull(val) && isFunction(val.then) && isFunction(val.catch)
-
-// Extra convenience named re-exports
-export { isArray, isObject, isPlainObject }
